requirements.txt
Flask==3.1.1
python-dotenv==1.1.0
openai==1.79.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug>=3.1.0
